+++ TestTools/StreamChatTestTools/Mocks/StreamChat/Workers/ChannelUpdater_Mock.swift
@@ -81,7 +81,7 @@ final class ChannelUpdater_Mock: ChannelUpdater {
     @Atomic var markUnread_userId: UserId?
     @Atomic var markUnread_messageId: MessageId?
     @Atomic var markUnread_lastReadMessageId: MessageId?
-    @Atomic var markUnread_completion: ((Error?) -> Void)?
+    @Atomic var markUnread_completion: ((Result<ChatChannel, Error>) -> Void)?
 
     @Atomic var enableSlowMode_cid: ChannelId?
     @Atomic var enableSlowMode_cooldownDuration: Int?
@@ -359,7 +359,7 @@ final class ChannelUpdater_Mock: ChannelUpdater {
         markRead_completion = completion
     }
 
-    override func markUnread(cid: ChannelId, userId: UserId, from messageId: MessageId, lastReadMessageId: MessageId?, completion: ((Error?) -> Void)? = nil) {
+    override func markUnread(cid: ChannelId, userId: UserId, from messageId: MessageId, lastReadMessageId: MessageId?, completion: ((Result<ChatChannel, Error>) -> Void)? = nil) {
         markUnread_cid = cid
         markUnread_userId = userId
         markUnread_messageId = messageId